corrected Gaia content
Gaia is measuring 1 billion stars (not 100 billion), and 1 billion
stars accounts for 1 percent or LESS (not 1 percent). The low-ball
estimation for stars in the Milky Way is 100 billion, but that number
is just averaging out the mass of the galaxy. The higher-end estimation
is 400 billion. Essentially, we don’t have a super accurate estimate of
the number of stars in our galaxy.